pyproject.toml
[tool.poetry]
name = "watersync"
version = "1.0.0"
description = "Hydrological data storage and analysis platform."
authors = ["Mateusz Zawadzki <zawadzkimat@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-slugify = "8.0.4"
Pillow = "10.4.0"
argon2-cffi = "23.1.0"
whitenoise = "6.7.0"
redis = "5.0.8"
hiredis = "3.0.0"
celery = "5.4.0"
django-celery-beat = "2.6.0"
flower = "2.0.1"
uvicorn = {extras = ["standard"], version = "0.30.6"}
uvicorn-worker = "0.2.0"
pandas = "2.2.2"
plotly = "5.23.0"
django = "5.0.8"
django-environ = "0.11.2"
django-model-utils = "4.5.1"
django-allauth = {extras = ["mfa"], version = "64.1.0"}
django-crispy-forms = "2.3"
crispy-bootstrap5 = "2024.2"
django-redis = "5.4.0"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
drf-spectacular = "0.27.2"
django-unfold = "0.36.0"
django-leaflet = "0.30.1"
django-jsonform = "2.22.0"
gunicorn = "23.0.0"
psycopg = {extras = ["c"], version = "3.2.1"}
sentry-sdk = "2.13.0"
django-anymail = {extras = ["sendgrid"], version = "11.1"}

[tool.poetry.group.docs.dependencies]


[tool.poetry.group.local.dependencies]
Werkzeug = "3.0.3"
ipdb = "0.13.13"
watchfiles = "0.21.0"
mypy = "1.11.1"
django-stubs = "5.0.4"
pytest = "8.3.2"
pytest-sugar = "1.0.0"
djangorestframework-stubs = "3.15.0"
ruff = "0.6.1"
coverage = "7.6.1"
djlint = "1.34.1"
pre-commit = "3.8.0"
factory-boy = "3.3.0"
django-debug-toolbar = "4.4.6"
django-extensions = "3.2.3"
django-coverage-plugin = "3.1.0"
pytest-django = "4.8.0"


# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["watersync/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
"COM812",
"ISC001",
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038",
]

[tool.ruff.lint.isort]
force-single-line = false